+++ semcore/data-table/src/components/Body/Body.types.ts
@@ -2,7 +2,7 @@ import type * as React from 'react';
 
 import type { CellPropsInner, Theme } from './Cell.types';
 import type { DTRow, UniqRowKey } from './Row.types';
-import type { DataRowItem, DTUse, VirtualScroll } from '../DataTable/DataTable.types';
+import type { DataRowItem, DTUse, VirtualScroll, DataTableProps } from '../DataTable/DataTable.types';
 import type { DTColumn } from '../Head/Column.types';
 
 export type CellRenderProps = {
@@ -68,7 +68,8 @@ export type BodyPropsInner = DataTableBodyProps & {
   getFixedStyle: (
     cell: Pick<DTColumn, 'name' | 'fixed'>,
   ) => [side: 'left' | 'right', style: string | number] | [side: undefined, style: undefined];
-  accordionDuration?: number | [number, number];
+  accordionDuration?: DataTableProps<any>['accordionDuration'];
   onCellClick: CellPropsInner['onClick'];
   rawData: DataRowItem[];
+  accordionMode?: DataTableProps<any>['accordionMode'];
 };